requirements.txt
fastf1==3.2.0
pandas==2.0.3
matplotlib==3.7.2
numpy==1.24.3
python-dotenv==1.0.0
supabase==1.2.0
requests==2.31.0
python-dateutil==2.8.2
tabulate==0.9.0 